requirements.txt
numpy>=1.26
pygame>=2.5
pytest
pytest-playwright
